push-notif-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sse-starlette
